+++ src/lib/utils/egeo-form/index.ts
@@ -1,0 +1,3 @@
+export { EgeoFormUtils } from './egeo-form-utils';
+export { StFormLabelModule, StFormLabelStatus } from './st-form-label';
+