+++ src/lib/utils/egeo-form/st-form-label/index.ts
@@ -1,0 +1,2 @@
+export * from './st-form-label.module';
+export * from './st-form-label-status.enum';